% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInfo.R
\name{addInfo}
\alias{addInfo}
\alias{addInfo.voteList}
\alias{addInfo.questionList}
\title{Add information on cabinets, elections and parties to dataset. This data
is obtained from ParlGov (http://www.parlgov.org/).}
\usage{
addInfo(x, ...)

\method{addInfo}{voteList}(x, ...)

\method{addInfo}{questionList}(x, ...)
}
\arguments{
\item{x}{A voteList object}

\item{...}{Other parameters passed on.}
}
\value{
A voteList object
}
\description{
Add information on cabinets, elections and parties to dataset. This data
is obtained from ParlGov (http://www.parlgov.org/).
}
\details{
This is a wrapper that runs both addCabinetInfo and addPartyInfo with default settings.
}
\section{Methods (by class)}{
\itemize{
\item \code{addInfo(voteList)}: Add information to voteList object

\item \code{addInfo(questionList)}: Add information to questionList object

}}
\examples{
examplevotes_with_info <- addInfo(examplevotes)
}
