% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosponsors.R
\name{cosponsors}
\alias{cosponsors}
\title{Calculate cosponsorship between MPs}
\usage{
cosponsors(voteList)
}
\arguments{
\item{voteList}{A voteList object, most of the time the votes objects that comes with the dutchparl package.}
}
\value{
A list with cosponsorship data in which each row represents a pair of MPs.
}
\description{
Calculate cosponsorship between MPs
}
\details{
Note that this operation can be expensive for very large datasets. It is recommended to limit the data to one parliamentary term.
}
\examples{
cosponsors(examplevotes)
}

