% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadQuestions}
\alias{downloadQuestions}
\title{Download Dutch Parliamentary Behaviour Questions Dataset from the internet}
\usage{
downloadQuestions(url = NULL, destfile = "questions.rds")
}
\arguments{
\item{url}{URL to download from. Defaults to downloading from dutchparl dataverse.}

\item{destfile}{Destination file (rds).}
}
\value{
Nothing. Side effect is to download the file.
}
\description{
Download Dutch Parliamentary Behaviour Questions Dataset from the internet
}
\examples{
\dontrun{
downloadQuestions()
}
}
