% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{examplequestions}
\alias{examplequestions}
\title{Dutch Parliamentary Behaviour Dataset (Questions)}
\format{A questionList object}
\usage{
data(examplequestions)
}
\description{
Written Questions from Members of the Dutch lower house 3-5 January 2010.
}
\keyword{datasets}
