% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{examplevotes}
\alias{examplevotes}
\title{Dutch Parliamentary Behaviour Dataset (Votes)}
\format{A voteList object}
\usage{
data(examplevotes)
}
\description{
Votes taken in the Dutch lower house between 1946-2015. See full documentation
for details on coding and use.
}
\keyword{datasets}

