% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{randomvotes}
\alias{randomvotes}
\title{Select a random number of votes from a voteList object}
\usage{
randomvotes(voteList, size = 10)
}
\arguments{
\item{voteList}{A voteList object}

\item{size}{Size of random selection.}
}
\value{
The subsetted voteList object.
}
\description{
Select a random number of votes from a voteList object
}
\examples{
randomvotes(examplevotes)
}

