% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{subset.questionList}
\alias{subset.questionList}
\title{Subset questionList object}
\usage{
\method{subset}{questionList}(x, df, subset, select, drop = FALSE,
  drop.levels = TRUE, ...)
}
\arguments{
\item{x}{A questionList object, most of the time the questions object from the Dutch Parliamentary Behaviour Dataset.}

\item{df}{The name of the data.frame in the questionList to filter on. Options include metaList, questionerList, responderList, and categoryList.}

\item{subset}{The subset command.}

\item{select}{Expression, indicating columns to select from data frame}

\item{drop}{passed on to [ indexing operator}

\item{drop.levels}{If true, superfluous levels in the data.frames will be removed.}

\item{...}{Other parameters (ignored)}
}
\value{
The subsetted questionList object.
}
\description{
Subset questionList object
}
\examples{
subset(examplequestions, examplequestions$metaList, dateQuestion > as.Date("2010-01-04"))
}
