% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{subset.voteList}
\alias{subset.voteList}
\title{Subset voteList object}
\usage{
\method{subset}{voteList}(x, df, subset, select, drop = FALSE,
  drop.levels = TRUE, ...)
}
\arguments{
\item{x}{A voteList object, most of the time the votes objects that comes with the dutchparl package.}

\item{df}{The name of the data.frame in the voteList to filter on. Options include metaList, sponsorList, and categoryList.}

\item{subset}{The subset command.}

\item{select}{Expression, indicating columns to select from data frame}

\item{drop}{passed on to [ indexing operator}

\item{drop.levels}{If true, superfluous levels in the data.frames will be removed.}

\item{...}{Other parameters (ignored)}
}
\value{
The subsetted voteList object.
}
\description{
Subset voteList object
}
\examples{
subset(examplevotes, examplevotes$metaList, date > as.Date("2010-01-15"))
}
