% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInfo.R
\name{addCabinetInfo}
\alias{addCabinetInfo}
\alias{addCabinetInfo.voteList}
\title{Add information on cabinet composition and elections to metadata. This data
is obtained from ParlGov (http://www.parlgov.org/).}
\usage{
addCabinetInfo(x, ...)

\method{addCabinetInfo}{voteList}(x, ...)
}
\arguments{
\item{x}{A voteList object}

\item{...}{Other parameters passed on.}
}
\value{
A voteList object
}
\description{
Add information on cabinet composition and elections to metadata. This data
is obtained from ParlGov (http://www.parlgov.org/).
}
\section{Methods (by class)}{
\itemize{
\item \code{voteList}: Cabinet Information for voteList object
}}
\examples{
addCabinetInfo(examplevotes)
}

