% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.rollcall.R
\name{as.rollcall}
\alias{as.rollcall}
\alias{as.rollcall.voteList}
\title{Transform into rollcall object}
\usage{
as.rollcall(x, ...)

\method{as.rollcall}{voteList}(x, ...)
}
\arguments{
\item{x}{Object to be transformed}

\item{...}{Other parameters (ignored)}
}
\value{
Rollcall object from package 'pscl'
}
\description{
Transform into rollcall object
}
\section{Methods (by class)}{
\itemize{
\item \code{voteList}: Transform voteList object into rollcall object
}}

\examples{
as.rollcall(examplevotes)
}
