% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadVotes}
\alias{downloadVotes}
\title{Download Dutch Parliamentary Behaviour Dataset from the internet}
\usage{
downloadVotes(url = NULL)
}
\arguments{
\item{url}{URL to download from. Defaults to downloading from dutchparl dataverse.}
}
\value{
voteList object
}
\description{
Download Dutch Parliamentary Behaviour Dataset from the internet
}
\examples{
\dontrun{
downloadVotes()
}
}

