% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{examplevotes}
\alias{examplevotes}
\alias{examplevotes}
\alias{examplequestions}
\title{Dutch Parliamentary Behaviour Dataset (Votes)}
\format{A voteList object}
\usage{
data(examplevotes)

data(examplequestions)
}
\description{
Votes taken in the Dutch lower house in January 2010.

Written Questions from Members of the Dutch lower house 3-5 January 2010.
}
\keyword{datasets}
