% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.R
\name{rice}
\alias{rice}
\alias{rice.voteList}
\title{Calculate rice score}
\usage{
rice(x, ...)

\method{rice}{voteList}(x, minvotes = 10, ...)
}
\arguments{
\item{x}{A voteList object}

\item{...}{Other parameters passed on.}

\item{minvotes}{The minimum number of votes for a party to have participated in. Defaults to 10.}
}
\value{
A list of rice scores.
}
\description{
Calculate rice score
}
\section{Methods (by class)}{
\itemize{
\item \code{voteList}: Rice index for voteList object
}}

\examples{
rice(examplevotes)
}
